update ready for ptod; with rest services

+++ src/main/java/ru/com/avs/controller/MainRestController.java
@@ -19,21 +19,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import ru.com.avs.model.Metal;
-import ru.com.avs.model.Mode;
-import ru.com.avs.model.Tare;
-import ru.com.avs.model.TempTare;
-import ru.com.avs.model.ThreadModel;
-import ru.com.avs.model.Waybill;
-import ru.com.avs.model.Weighing;
-import ru.com.avs.service.MetalService;
-import ru.com.avs.service.ModeService;
-import ru.com.avs.service.PropertyService;
-import ru.com.avs.service.TareService;
-import ru.com.avs.service.TempTareService;
-import ru.com.avs.service.ThreadService;
-import ru.com.avs.service.WaybillService;
-import ru.com.avs.service.WeighingService;
+import ru.com.avs.model.*;
+import ru.com.avs.service.*;
 import ru.com.avs.thread.CameraThread;
 import ru.com.avs.thread.ScaleThread;
 import ru.com.avs.util.SpringLoader;
@@ -64,6 +51,9 @@ public class MainRestController extends AbstractController {
 
     private ModeService modeService;
 
+    @Autowired
+    private CarService carService;
+
     /**
      * Send list of metals service to client.
      *
@@ -89,6 +79,14 @@ public class MainRestController extends AbstractController {
         ObjectMapper objectMapper = new ObjectMapper();
 
         return objectMapper.writeValueAsString(tares);
+    }
+
+    @RequestMapping(value = "/list-cars", method = RequestMethod.GET, produces = "application/json; charset=utf-8")
+    public String getListCars() throws IOException {
+        List<Car> cars = carService.getList();
+        ObjectMapper objectMapper = new ObjectMapper();
+        System.out.println("getcars");
+        return objectMapper.writeValueAsString(cars);
     }
 
     /**
@@ -163,11 +161,13 @@ public class MainRestController extends AbstractController {
 
         Weighing weighing = waybill.getWeighings().stream().findFirst().get();
         weighing.setWaybill(waybill);
+
         if (weighing.getBrutto().longValue() < weighing.getTare().longValue()) {
             BigDecimal tmp = weighing.getBrutto();
             weighing.setBrutto(weighing.getTare());
             weighing.setTare(tmp);
         }
+
         weighingService.save(weighing);
 
         threadService = (ThreadService) SpringLoader.getBean("ThreadService");
@@ -185,6 +185,15 @@ public class MainRestController extends AbstractController {
      *
      * @return ResponseEntity
      */
+    @RequestMapping(value = "/hello", method = RequestMethod.GET, produces = "application/json; charset=utf-8")
+    public ResponseEntity getHello() throws IOException {
+            System.out.println("\n\n\n\n\n\n HEllo");
+            String json = "hello";
+            return new ResponseEntity<>(json, HttpStatus.OK);
+    }
+
+
+
     @RequestMapping(value = "/get-settings", method = RequestMethod.POST, produces = "application/json; charset=utf-8")
     public ResponseEntity getSettings(@RequestParam(value = "modeId") Integer modeId) throws IOException {
         modeService = (ModeService) SpringLoader.getBean("ModeService");
@@ -198,6 +207,8 @@ public class MainRestController extends AbstractController {
             return new ResponseEntity<>("Не выбран режим", HttpStatus.BAD_REQUEST);
         }
     }
+
+
 
     /**
      * Delete weighing by id.
@@ -254,6 +265,28 @@ public class MainRestController extends AbstractController {
         return ResponseEntity.ok(retJson);
     }
 
+    @RequestMapping(value = "set-car", method = RequestMethod.PUT, produces = "application/json; charset=utf-8")
+    public ResponseEntity createTempCar(@RequestBody() String json) throws IOException {
+
+        ObjectMapper objectMapper = new ObjectMapper();
+        Waybill waybill = objectMapper.readValue(json, Waybill.class);
+
+        if (waybill.getId() == 0) {
+            waybill = createWaybillModel(waybill);
+            waybillService.save(waybill);
+        }
+
+        Car car = waybill.getCars().stream().findFirst().get();
+        car.setWaybill(waybill);
+        carService.save(car);
+
+        waybill = waybillService.getById(waybill.getId());
+        List<Car> carList = new ArrayList<>();
+        carList.add(car);
+        waybill.setCars(carList);
+        String retJson = objectMapper.writeValueAsString(waybill);
+        return ResponseEntity.ok(retJson);
+    }
     /**
      * Delete TempTare.
      *
@@ -277,7 +310,7 @@ public class MainRestController extends AbstractController {
     public ResponseEntity stop(@RequestParam(value = "waybillId") Integer waybillId) {
         Waybill waybill = waybillService.getById(waybillId);
 
-        if (waybill.getWeighings().size() == 0 && waybill.getTempTares().size() == 0) {
+        if (waybill.getWeighings().size() == 0 && waybill.getTempTares().size() == 0 && waybill.getCars().size() == 0) {
             waybillService.delete(waybill);
         }
         return ResponseEntity.ok("ok");
